requirements.txt
certifi==2025.4.26
charset-normalizer==3.4.2
colorama==0.4.6
contourpy==1.1.1
cycler==0.12.1
einops==0.6.1
fonttools==4.57.0
idna==3.10
importlib_resources==6.4.5
kiwisolver==1.4.7
matplotlib==3.7.2
numpy==1.23.5
opencv-python==4.8.0.74
packaging==25.0
pillow==10.4.0
pyparsing==3.0.9
python-dateutil==2.9.0.post0
PyYAML==6.0
requests==2.32.3
six==1.17.0
torch==1.11.0
torchvision==0.12.0
tqdm==4.65.0
typing_extensions==4.13.2
urllib3==2.2.3
zipp==3.20.2
